Remove Hazelcast dependency
The only remaining use was a testing REST backend in the LRE.

I wrote a replacement for that using netty, which we use for our network stuff in Bazel, which means we can now get rid of Hazelcast. :)

I'll remove the Hazelcast files in a separate change when this is merged.

PiperOrigin-RevId: 205985996