Order input files from TreeArtifacts
This commit fixes non-deterministic behaviour when it comes to the usage
of C++ TreeArtifacts for compilation / linkage.

If a folder is passed to a cc_library rule, the source / header files
are discovered at execution phases. The order of which those files
are served by the OS is not be deterministic. This might lead to
non reproducable builds, meaning the order of each object file inside
the created library might differ.
This can be fixed by using a SortedMap instead of Maps.

This commit might have side-effects to other rules that also utilize
TreeArtifacts, such that the inputs will be sorted here also.