Automated rollback of commit 7e87730de985b7099b9b683571d58efdaab70890.
*** Reason for rollback ***

Go back to the default constructor - instead of requiring everywhere to know the correct hash function, we'll have the default rely on global state. It will make transition easier, even if it makes the origin of the hash less obvious.

*** Original change description ***

Remove default MD5 in most of Bazel's virtual filesystems.

This forces the ex-default to be explicit in a lot of tests, but I'd rather that than have the risk of implicit md5-use in production code.

To keep this CL smaller, do not remove the default from UnixFS quite yet.

RELNOTES: None.
PiperOrigin-RevId: 206358838