Add aarch64 embedded JDK config, MacOS mock toolchain settings
Hi,

This PR is for additional work to get the tests running correctly on Aarch64. One key issue was that the default embedded jvm used by default was X86_64 and does not work on other architectures.
To support this I have added a new constraint, however I will refine this in case a version of Darwin becomes available on aarch64.

I'll do the work to find a good arm32 toolchain in a future CL. I do not have access for ppc or z390 to repair those architectures.

Either way, hope this is good for you.

Closes #5525.

PiperOrigin-RevId: 206717726