Slightly improve UnixGlob implementation. For recursive patterns, the current implementation leads to two identical recursive calls. This isn't harmful as there is a cache that will make the second call very cheap, but does not seem right.
As a consequence, a cache is only needed if there are two recursive patterns in
a single glob-path, e.g. /a/b/**/c/**/d/... as either "**" can expand to 0 or
more directories (so /a/b/c/c/d could be reached in two different ways). If
there is only a single recursive pattern, even with "*" placeholders, there is
always a unique expansion (a "*" always represents exactly one path element).

RELNOTES: None.
PiperOrigin-RevId: 206753919