Change Deque -> LinkedHashSet
In some cases, we may be reporting a lot of actions, and this has shown up as a hotspot in our profiling. LinkedHashSet should be just as fast for insertion, keep order as Deque, and be faster for removal.

There are a couple of more aggressive steps we can take if this doesn't help.

PiperOrigin-RevId: 207064758