[Skylark] Speed up string.partition function.
According to JMH using `ImmutableList.of` or `Arrays.asList` is ~2X faster
than using existing approach of creating an empty `ArrayList` with expected
size and populating it using `add` method. This is most likely due to extra
range and capacity checks.

Returning `ImmutableList` instead of `ArrayList` avoids the need to copy it
again in order to create a `SkylarkTuple`.

These changes speed up the function by ~3X.

Closes #5736.

PiperOrigin-RevId: 207079605