Automated rollback of commit a6255612e4892729d3758775c76085b26b9bc584.
*** Reason for rollback ***

Breaks FDO builds, see http://b/112112125

*** Original change description ***

Remove PerActionFileCache

Instead, make ActionMetadataHandler implement the MetadataProvider interface.

This fixes an issue where an action that runs two spawns where one depends on
an output of the other was unable to get the metadata for the intermediate
output.

We don't currently have actions that do this, but we will have in a future
change (which will also implicitly act as a regression test).

PiperOrigin-RevId: 207084179