remote: add support for directory inputs in runfiles
Add support for tree artifacts (ctx.action.declare_directory(...)) in
runfiles. Before this change we would throw away the information
about the files inside a tree artifact before executing an action.
That's fine for local execution where the sandbox just copies/symlinks
a directory and doesn't care much what's inside. However, in remote
execution we actually need to upload each individual file and so
we need to be aware of all individual files not just directories.

This change makes it so that this information is made available to a
SpawnRunner via the SpawnInputExpander.

RELNOTES: None
PiperOrigin-RevId: 207091668