Clarifies the documentation for action.args() set_param_file_format.
The way documentation reads, it sounds like 'shell' format is single line (especially because the other format is called multiline).
Currently it seems like the implementation only uses new line but assuming Bazel doesn't want to promise on that, this patch updates the documentation to point that whitespace could be space, tab or newline - which mimics the documentation of ParameterFileType.SHELL_QUOTED.

RELNOTES: None.
PiperOrigin-RevId: 207142679