Remove non-static ApplicationManifest methods
Remove all state from the ApplicationManifest class. Actual methods in this
class will be migrated to somewhere more appropriate in upcoming changes.

We now usually use the manifest passed around in the ResourceApk object
instead. (The exception is mobile-install, which still wants to use the raw
manifest.) Note that this manifest is the processed manifest output by resource
processing - using it will somewhat decrease parallelisim, but is also more
correct. (We'll get the parallelisim gains back once we now longer use the
legacy manifest merger and don't need to process the manifest in resource
processing.)

RELNOTES: none
PiperOrigin-RevId: 207881830