Create BUILD.tools for //tools/test package.
Motivation: as part of fixing issue #5508, I'll
add a cc_binary to this package. The binary will
be embedded in Bazel as a prebuilt binary (to
avoid requiring a C++ compiler to use it).
Therefore //tools/test/BUILD and
@bazel_tools//tools/test/BUILD will contain
different rules, mandating the BUILD file split.

See https://github.com/bazelbuild/bazel/issues/5508

Change-Id: If21bafbc3d83d59e52de498cf3efd2b1b79fa7b6

Closes #5787.

Change-Id: If21bafbc3d83d59e52de498cf3efd2b1b79fa7b6
PiperOrigin-RevId: 208005382