Fix StackOverflowError on Windows. Fixes #5730
We found that with JDK9 and up Bazel would sometimes crash
with a StackOverflowError in one of the Command-Accumulator-Thread-*
threads. We experimentally found that this error was due to these
threads being constrained to a 32KiB stack size. The default stack
size for JVM threads on most 64-bit systems is 1MiB (So that's 3%
of the default). The purpose of the Command-Accumulator-Threads
is to read stdout/stderr from processes that Bazel launches locally.

The proposed fix is to just use the system default stack size for
these threads. The alternative is to increase the size limit to
some arbitrary number that happens to work, but this is likely
premature optimization and I'd like to avoid that if possible. We
further found that this code even predates Blaze/Bazel and is
from 2005.

PiperOrigin-RevId: 208009940