Support a .bazelignore file
Add suport for a file .bazelignore, allowing to specify directories
that should be ignored. So far, even for the "ignored" directories,
the directory structure, including following symlinks, still has
to be well founded. We plan to lift that restriction in the future,
but already with the restriction, it is an improvement over the
status quo.

RELNOTES: bazel now supports a .bazelignore file specifying
  directories to be ignored; however, these directories still
  have to be well founded and, in particular, may not contain
  symlink cycles.

Improves on #4888.

Change-Id: Icfe0b7f5c35eceea16e28944ca7711431b71271a
PiperOrigin-RevId: 208050372