Reduce contention in the digest map.
During reads and writes a read lock is taken. If we need to resize the table, a write lock is taken to flush all the readers and writers before the table is resized.

The read lock should be significantly cheaper than full synchronisation since multiple threads can grab a read lock at once.

RELNOTES: None
PiperOrigin-RevId: 208059810