Introduce a retriever for logging handler properties.
We want a way for Bazel to find a logging handler's current log file without
direct dependencies on the exact handler class. We do this with an abstract
parent class whose concrete child class (to be used as a singleton) will be
given in startup_options, i.e. in the same place as the server logging
configuration.

RELNOTES: None.
PiperOrigin-RevId: 208171084