Rename target to avoid cycle.
The cc_binary //tools/test:test_wrapper produced
the output file "test_wrapper", then we put this
in "@bazel_tools//tools/test" and created the
filegroup "@bazel_tools//tools/test:test-wrapper"
to contain it. The filegroup and the file had the
same name so Bazel thought it was a dependency
cycle.

Change-Id: Ic6742eaa3debfc1bc6f0399d2475f2575b7a4a5a