[pkg_tar] Package py_binary runfiles correctly
For py_binary rules it is necessary that the packaging of the runfiles
is done correctly, since otherwise the py_binary might not be able to
be executed, since dependencies are missing.

Currently the runfiles are not packaged in a according <target>.runfiles
directory, which leads undefined depdencies.

This commit creats such a <target>.runfiles directory as done by bazel
in the bazel-bin folder structure.

In order to ensure that other rules e.g. cc_binary, are not affected,
we check if our rule is python related.

Part of issue: #4383