Fix MultisetSemaphore.
We use a fixed version of the previous algorithm. See the comments for details.

Fancier algorithms exist. I thought of a cool one that makes use of BatchKeyedLocker (would give me an excuse to revive it, heh), but fancy algorithms would be overkill. As noted in the initial commit of NaiveMultisetSemaphore, performance isn't critical.

RELNOTES: None
PiperOrigin-RevId: 208560559