Add --filter_sources flag to LcovMerger.
When the `--filter_sources` flag is used `LcovMerger`  excludes from the merged coverage the sources whose names match any of the regex in the flag.

The value of `--filter_sources` is a list of regex separated by `,`.

This flag comes as a preparation for using `gcov` instead of `lcov` for collecting coverage data for C++. The flag was not needed before because `lcov` has some functionality for excluding some files (e.g. `--no-external` to ignore coverage data for system files).

Closes #5834.

PiperOrigin-RevId: 208606867