Change the name of the OptionsProvider class -> OptionsParsingResult. This is CL 1/3 in order to add a <code>getSkylarkOptions()</code> method to the top level OptionsProvider which is currently known as OptionsClassProvider.
OptionsClassProvider will be renamed to OptionsProvider in CL 2/3 in this series to better reflect the fact that the class will be a more general options provider instead of specifically an options ~class~ provider. This requires changing the current OptionsProvider class to something else, which is what this CL is doing.

PiperOrigin-RevId: 208885826