Automated rollback of commit 51215bd922d57f2ddd03deac6295f3e9267c7c3c.
*** Reason for rollback ***

This CL is breaking //devtools/blaze/integration:perf_regtest_rabbit guitar test because there is currently no way to resolve the new 'config' built-in with pre-existing config global variables created by users. Rolling back to unblock guitar for now. The solution is to allow global vars to shadow built-ins (b/35984389).

*** Original change description ***

Create new config module that is responsible for creating BuildSetting descriptor objects.

Build settings are units of configuration i.e. a key/value pair of a setting (e.g. cpu) and a value (e.g. ppc). A build setting descriptor is used to describe what kind of build setting a skylark rule is (if any at all). The BuildSettingDescriptor implementation of the API describes two facets of the build setting rule: the type of the value and whether or not the setting is settable on the command line....

***

ROLLBACK_OF=208669663
PiperOrigin-RevId: 208971834