Fix race condition with releaseUploadChannel().
A race condition can occur in HttpBlobStore.releaseUploadChannel() when
ChannelHandlerContext is closed by HttpUploadHandler.channelRead0. Because
ChannelHandlerContext.close runs asynchronously and closes all pipelines in
reverse order (before the channel is marked as closed),
HttpBlobStore.releaseUploadChannel() can be executed while channel is still
open and handlers are being removed, causing an exception to be thrown by the
ch.pipeline().remove(X) calls. This fix saves the status of the keepAlive
response from the HttpUploadHandler so that the channel can be fully closed
before releasing.

Resolves #5952