Add parametrized filesystem constructors that allow the caller to control the digest hash function.
Add a default value constant for use in tests so that tests don't need a call to setDefault() or a fallback in the FileSystem default constructor.
Followup changes will

(1) set this default in all tests,
(2) make InMemoryFileSystem, which is basically only used in tests, never call its superclass's constructor which uses the global default,
(3) remove the fallback so that the default constructor can throw DefaultNotSetException.

Splitting these changes apart for ease of review.

RELNOTES: None.
PiperOrigin-RevId: 209198187