HttpUtils: short cut in mergeUrls, if possible
While, in theory, we can decompose a URI into its components and recompose
it later this will give us an equivalent URI, but not necessarily the same:
the quoting might be different. Unfortunately, some sites rely on the precise
representation of a URL in their authentication scheme of a redirect URL.
So preserve the original representation whenever we do not have to inherit
anything from the original URL.

Fixes #5932.

Change-Id: I6901e044c405eefd3e627433f141e02dbce84b4a
PiperOrigin-RevId: 209415592