Unwrap CacheNotFound for RemoteActionCache
Using retrier causes CacheNotFoundException to be wrapped in a
RetryException, which is handled through undesirable means by the users
of RemoteActionCache#download via IOException. This must be unwrapped
and thrown directly as a part of the download interface for users, so
that users can distinguish these degraded cache entries specifically.

Reverting the behavior of remotelyReExecuteOrphanedCachedActions, which
was adapted to verify download retries, and should be separated into its
own test.

Closes #5917.

PiperOrigin-RevId: 209638152