Do not create the FDO root anymore.
Now that we don't have FdoStubAction, it's not necessary; it was used as a special root that's somewhere between a source and a derived root and it was all around awful.

Note that this changes the value of fdo_profile_path such that it is not set anymore in VANILLA mode but that mode is not supported anymore so that's okay.

RELNOTES: None.
PiperOrigin-RevId: 209926721