Change PlatformConfiguration to only have a single target platform.
This is the actual situation that has existed and will continue to be supported: each ConfiguredTarget has a single target platform. When we update to support multi-platform builds, the configuration will be split at the top level so that each split still only sees a single target platform.

Multi-arch platforms (such as fat binaries) will need a single platform to represent the multi-arch situation, and rules will need to know how to handle generating and combining the proper single-target dependencies: this is the same situation we have currently.

Change-Id: I904d74d136c761fa9c8da329040dd20920db00b7

Closes #5755.

Change-Id: Icaa539fa864d005c754986f7c98dac1bb8dc7e35
PiperOrigin-RevId: 209939618