Change Artifact.equals() to early-out if hashcodes differ.
java.util.HashMap.get() has an optimization where it compares the
hashcode to a cached value before calling equals() on the candidate key.
CompactHashSet/Map keep a table of cached hashcodes.
But RegularImmutableMap only has a comment saying it assumes the objects in
the collection will optimize equals() with == and hashcode as appropriate!

Simplify use of Objects.equal(); none of the fields are nullable.

Take getClass() out of hashcode computation.  I think two different
Artifact subclasses for the same path is unusual in normal operation,
and I may try to exploit the pass-through hashcode in ActionInputMap.

RELNOTES: None.
PiperOrigin-RevId: 210117899