hide warning on macOS if no JDK is installed
On macOS Bazel calls /usr/libexec/java_home to detect the
system jdk (if any). This tool prints a warning if no system
jdk is installed:

> Unable to find any JVMs matching version "1.7+".
> No Java runtime present, try --request to install.

This patch hides this warning as Bazel ships with an embedded
JDK and does not require a system JDK installed. Additionally,
pump the version requirement to 1.8+ as this is the oldest JDK
version that Bazel can build for.