Remove dead code in TestSummary.
Because TestCase is a protobuf, the getters can never return null for String types. If the field is not set, it will return the default for the type, which is the empty string. (cf. https://developers.google.com/protocol-buffers/docs/proto#optional)

The "correct" way to fix this would be to replace the null check with testCase.hasName() and testCase.hasClassName(), but by now the code hasn't worked like that for 2 years, so "fixing" it will probably do more damage by stuff invisibly relying on current behavior than simply deleting this.

RELNOTES: None.
PiperOrigin-RevId: 210361495