Print further details on an AbruptExitException.
Otherwise we just get a trivial message with the summary of the exception
that caused the failure... but zero context even if it exists.  This is a
requirement to get useful information when we fail hard.  Dumping a stack
trace may not be the nicest for the user, but given that this is improving
a catch-all location, I'd rather get more information than nothing.  If we
need to print more succinct messages in other circumstances, we can
special-case those.

While here, also add missing trailing newlines to a few messages that need
them.

RELNOTES: None.
PiperOrigin-RevId: 210409135