C++: Removes calls to setCcLinkparamsStore of CcLinkingInfo.Builder
This is in preparation for deleting CcLinkParamsStore. All remaining calls
to the setCcLinkparamsStore method of the CcLinkingInfo builder have been removed.

See b/111781390 for details of rollback.

Second rollback due to b/112656426. The reason for this one single failure not showing up in TGP was that I was changing the order in which we iterated over CcLinkingInfo providers. The order of dependencies in C++ targets is important. I have changed the PyBinary test to catch this.

RELNOTES:none
PiperOrigin-RevId: 210521734