Add missing tests for actions whose inputs do not change
The testActionWithNonChangingInput test never ran, because it was
parameterized and was missing @Test. The intent was to assert that
actions do not reexecute when their inputs are not dirtied, regardless
of whether the action is configured to "executeUnconditionally" or has
"isVolatile" equal to true.

Similar neighboring tests assert action behavior under different
circumstances, like when its input artifact's mtime or mtime-and-content
change.

RELNOTES: None.
PiperOrigin-RevId: 210744214