Add --unlimit_coredumps to raise the soft coredump limit to the hard limit.
This is a client option that, when given, attempts to raise the soft
coredump limit to the hard limit both for the client and the server.
This is useful because some systems nowadays (stock macOS, and our own
internal Linux installation) disable coredumps by default by setting the
soft limit to zero, but do not forbid the user from explicitly enabling
them (the hard limit is set to unlimited).  With this option, we can
configure Bazel once to support coredumps and forget about it.

I've been trying to get information from a very infrequent JVM crash that
Bazel is able to trigger on macOS but have been unable to do so with any
coredump data.  With this flag, I hope to be able to get more details.

I could create a custom wrapper script for the bazel binary that tweaks
the soft limit before invoking bazel... but having to carry such wrapper
across machines is annoying.  And given that stock configurations come
with soft limits set to zero, I think it's reasonable to have this as an
option within Bazel itself.

TESTED: Manually by running a Bazel build with this new flag turned on,
killing the server and the client, and seeing cores being generated.
I tried writing an automated integration test for this but it's way too
complex and fragile (e.g. the machine we run on may not even allow core
dumps or they may actually be limited in size, so there is not much the
test can do under those conditions).

RELNOTES: None.
PiperOrigin-RevId: 210967964