Environment: ensure that the global frame does at most two lookups
When we try to access a variable, we do a lookup in the environment. The globalframe has a parent field, so we have no limit on the number of lookups. It may do multiple lookups until it finds a value or the parent is null.

With this CL:
- globalframe represents the symbols from the file
- globalframe.parent represents the universe (builtins)
- globalframe.parent.parent won't exist

So it's at most two lookups. Later, we should reduce to just one lookup. After that, we can have further optimizations (e.g. array lookup instead of hashmap lookup).

https://github.com/bazelbuild/bazel/issues/5637

RELNOTES: None.
PiperOrigin-RevId: 211143657