[Remote] Do not allow concurrent uploads of the same digest
Resolves https://github.com/bazelbuild/bazel/issues/4664

This issue is reproducible locally without another user concurrently building and pushing to the remote cache. We have a medium-large C++ project with many targets with shared dependencies. With this patch, I no longer see

```
WARNING: Error writing to the remote cache:
429 Too Many Requests
<?xml version='1.0' encoding='UTF-8'?><Error><Code>SlowDown</Code><Message>Please reduce your request rate.</Message><Details>Changing object butterfly-software-bazel-cache/cas/fc503dd74d0add4a88a492388ee56f46e71ecd117dbac8177c0846d2d44f5221 too quickly, please slow down</Details></Error>
```

And as a result, a `build` -> `clean --expunge` -> `build` now gets 100% cache hits and prints no warnings.

Since this warning is concerned with changing an object too often, we could save some memory by limiting the size of the set and evicting the oldest entries.

Closes #6040.

PiperOrigin-RevId: 211328027